+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod common;
-pub mod config;
 mod user;
 
 use std::process::{Command, Stdio};